// src/shared/enums/variable-environment.enum.ts
export enum VARIABLE_ENVIRONMENT_ENUM {
  JWT_ACCESS_TOKEN_SECRET = 'JWT_ACCESS_TOKEN_SECRET',
  JWT_ACCESS_TOKEN_EXPIRES_IN = 'JWT_ACCESS_TOKEN_EXPIRES_IN',
  JWT_REFRESH_TOKEN_SECRET = 'JWT_REFRESH_TOKEN_SECRET',
  JWT_REFRESH_TOKEN_EXPIRES_IN = 'JWT_REFRESH_TOKEN_EXPIRES_IN',
  JWT_SPECIAL_TOKEN_SECRET = 'JWT_SPECIAL_TOKEN_SECRET',
  JWT_SPECIAL_TOKEN_EXPIRES_IN = 'JWT_SPECIAL_TOKEN_EXPIRES_IN',
  TWILIO_ACCOUNT_SID = 'TWILIO_ACCOUNT_SID',
  TWILIO_AUTH_TOKEN = 'TWILIO_AUTH_TOKEN',
  TWILIO_PHONE_NUMBER = 'TWILIO_PHONE_NUMBER',
  DATABASE_URL = 'DATABASE_URL',
  PORT = 'PORT',
}